+++ commands/update.go
@@ -23,20 +23,10 @@
 package commands
 
 import (
-	"bufio"
-	"bytes"
 	"fmt"
-	"io"
-	"net/http"
-	"os"
-	"os/exec"
-	"runtime"
 	"strings"
 
-	browser "github.com/EDDYCJY/fake-useragent"
-	"github.com/alecthomas/chroma/quick"
 	"github.com/mkloubert/go-package-manager/types"
-	"github.com/mkloubert/go-package-manager/utils"
 	"github.com/spf13/cobra"
 )
 
@@ -51,215 +41,45 @@ func Init_Update_Command(parentCmd *cobra.Command, app *types.AppContext) {
 	var userAgent string
 
 	var updateCmd = &cobra.Command{
-		Use:     "update",
+		Use:     "update <modules>",
 		Aliases: []string{"upd"},
 		Short:   "Update dependencies",
-		Long:    `Updates all dependencies in this project.`,
+		Long:    `Updates all or only specific dependencies in this project.`,
 		Run: func(cmd *cobra.Command, args []string) {
 			if selfUpdate {
-				app.Debug("Will start self-update ...")
-
-				consoleFormatter := utils.GetBestChromaFormatterName()
-				consoleStyle := utils.GetBestChromaStyleName()
-
-				customUserAgent := strings.TrimSpace(userAgent)
-				if customUserAgent == "" {
-					customUserAgent = browser.Chrome()
-				}
-
-				customPowerShellBin := strings.TrimSpace(powerShellBin)
-				if customPowerShellBin == "" {
-					customPowerShellBin = "powershell"
-				}
-
-				customUpdateScript := strings.TrimSpace(updateScript)
-				if customUpdateScript == "" {
-					customUpdateScript = strings.TrimSpace(os.Getenv("GPM_UPDATE_SCRIPT"))
-				}
-
-				downloadScript := func(url string) ([]byte, error) {
-					app.Debug(fmt.Sprintf("Download from '%s' ...", url))
-					app.Debug(fmt.Sprintf("User agent: %s", customUserAgent))
-
-					req, err := http.NewRequest("GET", url, bytes.NewBuffer([]byte{}))
-					if err != nil {
-						return []byte{}, err
-					}
-
-					req.Header.Set("User-Agent", customUserAgent)
-
-					client := &http.Client{}
-					resp, err := client.Do(req)
-					if err != nil {
-						return []byte{}, err
-					}
-					defer resp.Body.Close()
-
-					if resp.StatusCode != 200 {
-						return []byte{}, fmt.Errorf("unexpected response: %v", resp.StatusCode)
-					}
-
-					responseData, err := io.ReadAll(resp.Body)
-
-					return responseData, err
-				}
-
-				showNewVersion := func() {
-					if noVersionPrint {
-						return
-					}
-
-					app.RunShellCommandByArgs("gpm", "--version")
-				}
-
-				if powerShell || utils.IsWindows() {
-					// PowerShell
-					app.Debug(fmt.Sprintf("Will use PowerShell '%s' ...", powerShellBin))
-
-					scriptUrl := customUpdateScript
-					if scriptUrl == "" {
-						scriptUrl = "https://raw.githubusercontent.com/mkloubert/go-package-manager/main/sh.kloubert.dev/gpm.ps1"
-					} else {
-						su, err := utils.ToUrlForOpenHandler(scriptUrl)
-						utils.CheckForError(err)
-
-						scriptUrl = su
-					}
-
-					pwshScript, err := downloadScript(scriptUrl)
-					utils.CheckForError(err)
-
-					executeScript := func() {
-						p := exec.Command(customPowerShellBin, "-NoProfile", "-Command", "-")
-						p.Dir = app.Cwd
-						p.Stderr = os.Stderr
-						p.Stdout = os.Stdout
-
-						stdinPipe, err := p.StdinPipe()
-						utils.CheckForError(err)
-
-						err = p.Start()
-						utils.CheckForError(err)
-
-						go func() {
-							defer stdinPipe.Close()
-							stdinPipe.Write([]byte(pwshScript))
-						}()
-
-						err = p.Wait()
-						utils.CheckForError(err)
-
-						showNewVersion()
-						os.Exit(0)
-					}
-
-					if force {
-						executeScript()
-					} else {
-						// ask the user first
-
-						err = quick.Highlight(os.Stdout, string(pwshScript), "powershell", consoleFormatter, consoleStyle)
-						if err != nil {
-							fmt.Print(string(pwshScript))
-						}
-
-						fmt.Println()
-						fmt.Println()
-
-						reader := bufio.NewReader(os.Stdin)
-
-						for {
-							fmt.Print("Do you really want to run this PowerShell script (Y/n)? ")
-							userInput, _ := reader.ReadString('\n')
-							userInput = strings.TrimSpace(strings.ToLower(userInput))
-
-							switch userInput {
-							case "", "y", "yes":
-								executeScript()
-							case "n", "no":
-								os.Exit(0)
-							}
-						}
-					}
-				} else if utils.IsPOSIXLikeOS() {
-					// if POSIX-like => sh
-					app.Debug("Will use UNIX shell ...")
-
-					scriptUrl := customUpdateScript
-					if scriptUrl == "" {
-						scriptUrl = "https://raw.githubusercontent.com/mkloubert/go-package-manager/main/sh.kloubert.dev/gpm.sh"
-					} else {
-						su, err := utils.ToUrlForOpenHandler(scriptUrl)
-						utils.CheckForError(err)
-
-						scriptUrl = su
-					}
-
-					bashScript, err := downloadScript(scriptUrl)
-					utils.CheckForError(err)
-
-					executeScript := func() {
-						p := exec.Command("sh")
-						p.Dir = app.Cwd
-						p.Stderr = os.Stderr
-						p.Stdout = os.Stdout
-
-						stdinPipe, err := p.StdinPipe()
-						utils.CheckForError(err)
-
-						err = p.Start()
-						utils.CheckForError(err)
-
-						go func() {
-							defer stdinPipe.Close()
-							stdinPipe.Write([]byte(bashScript))
-						}()
-
-						err = p.Wait()
-						utils.CheckForError(err)
-
-						showNewVersion()
-						os.Exit(0)
-					}
-
-					if force {
-						executeScript()
-					} else {
-						// ask the user first
-
-						err = quick.Highlight(os.Stdout, string(bashScript), "shell", consoleFormatter, consoleStyle)
-						if err != nil {
-							fmt.Print(string(bashScript))
-						}
-
-						fmt.Println()
-						fmt.Println()
-
-						reader := bufio.NewReader(os.Stdin)
-
-						for {
-							fmt.Print("Do you really want to run this bash script (Y/n)? ")
-							userInput, _ := reader.ReadString('\n')
-							userInput = strings.TrimSpace(strings.ToLower(userInput))
-
-							switch userInput {
-							case "", "y", "yes":
-								executeScript()
-							case "n", "no":
-								os.Exit(0)
-							}
-						}
-					}
-				} else {
-					utils.CheckForError(fmt.Errorf("self-update for %s/%s is not supported yet", runtime.GOOS, runtime.GOARCH))
-				}
+				run_self_update_command(
+					app,
+					force, noVersionPrint, powerShell, powerShellBin, updateScript, userAgent,
+				)
 			} else {
-				app.Debug("Will start project dependencies ...")
+				modulesToUpdate := make([]string, 0)
+				for _, moduleNameOrUrl := range args {
+					// maybe alias => module url(s)
+					moduleUrls := app.GetModuleUrls(moduleNameOrUrl)
 
-				app.RunShellCommandByArgs("go", "get", "-u", "./...")
+					modulesToUpdate = append(modulesToUpdate, moduleUrls...)
+				}
+
+				additionalShellArgs := make([]string, 0)
+				additionalShellArgs = append(additionalShellArgs, modulesToUpdate...)
+				if len(modulesToUpdate) == 0 {
+					app.Debug("Will update all modules in project ...")
+
+					// update all in this project instead specific onces
+					additionalShellArgs = append(additionalShellArgs, "./...")
+				} else {
+					// update specific ones
+					app.Debug(fmt.Sprintf("Will update following modules in project: %s", strings.Join(modulesToUpdate, ",")))
+				}
+
+				allShellArgs := make([]string, 0)
+				allShellArgs = append(allShellArgs, "get", "-u")
+				allShellArgs = append(allShellArgs, additionalShellArgs...)
+
+				app.RunShellCommandByArgs("go", allShellArgs...)
 
 				if !noCleanup {
-					app.RunShellCommandByArgs("go", "mod", "tidy")
+					app.TidyUp()
 				}
 			}
 		},